+++ src/kernel/cmdif_srv_flib.c
@@ -297,7 +297,7 @@ __COLD_CODE int cmdif_srv_open(void *_srv,
 	void   *dev = NULL;
 
 #ifdef DEBUG
-	if (v_data & 0x7)
+	if ((uint64_t)v_data & 0x7)
 		return -EINVAL;
 
 	if ((v_data != NULL) && (size < CMDIF_SESSION_OPEN_SIZEOF))
